PS-9673 fix of bug with double memory freeing in case of an error
https://perconadev.atlassian.net/browse/PS-9673

The problem: encoding buffer address is referenced in 2 places, in the local variable
and in the ctx->buffer. When buffer was dealocated with "kmip_free_buffer(&ctx, encoding, buffer_total_size)" call,
and then kmip_destroy(&ctx) was called, we had double deallocation.
To fix the problem, "kmip_set_buffer(&ctx, NULL, 0)" call was added before context destruction call in all places.
There's no such problem in "happy path" because "kmip_set_buffer(&ctx, NULL, 0)" is there.

This is not the ideal solution, because whole "kmip_bio.c" is a messy code that should be re-written ASAP.